new data download files

+++ DataDownload/NIMROD/Download_from_CEDA.py
@@ -6,24 +6,19 @@ import tarfile
 import gzip
 import shutil
 
-
-# year = int(sys.argv[1])
-year = 2018
+year = int(sys.argv[1])
+# 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020
 
 # When script is run this allows user to type username and password without showing it
-uuname = input('username:')
-ppword = getpass('password:')
-
-# uuname = sys.argv[1] #'masher'
-# ppword = 'Ug@nda!1'
-
+uuname = 'masher'
+ppword = 'T{$WW"8bLlAt'
 
 # Login to FTP
-f=ftplib.FTP("anon.ftp.ceda.ac.uk", uuname, ppword)
-
+f = ftplib.FTP("ftp.ceda.ac.uk", uuname, ppword)
 
 # Loop through years 
 for year in range(year, year+1, 1):
+    print(year)
     ###########################
     # Change to right directory
     ###########################
@@ -51,16 +46,20 @@ for year in range(year, year+1, 1):
     # Loop through list and download files
     ###########################
     for ffile in dir_list:
-        if not ffile[35:37] in ['01', '02', '03', '04', '05', '09', '10', '11','12']:
-#         if ffile == 'metoffice-c-band-rain-radar_uk_20040831_1km-composite.dat.gz.tar':
-            with open(ffile, 'wb') as fp:
-                print("Downloading")
-                # Download the file
-                # If the file doesn't exist then this causes the exception to be generated
-                res = f.retrbinary('RETR %s' % ffile , fp.write)
-                # Don't know why this bit was there.
-                if not res.startswith('226 Transfer complete'):
-                    print(ffile, 'Download failed')
-                    if os.path.isfile(ffile):
-                        os.remove(ffile)         
+#         if ffile[35:37] in ['06', '07', '08']:
+#             print(ffile[37:39])
+#             if ffile[37:39] in ['26']:
+        if ffile == 'metoffice-c-band-rain-radar_uk_20121117_1km-composite.dat.gz.tar':
+                with open(ffile, 'wb') as fp:
+                    print("Downloading")
+                    # Download the file
+                    # If the file doesn't exist then this causes the exception to be generated
+                    res = f.retrbinary('RETR %s' % ffile , fp.write)
+                    # Don't know why this bit was there.
+                    if not res.startswith('226 Transfer complete'):
+                        print(ffile, 'Download failed')
+                        if os.path.isfile(ffile):
+                            os.remove(ffile)         
     print(f"Download complete for {year}")
+    
+
